app/requirements.txt
Django
python-decouple
graphene-django
django-graphql-jwt
celery
redis
django-redis
flower
psycopg2-binary